requirements.txt
# Discord Bot Dependencies
discord.py>=2.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.0

# RAG System Dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
openai>=1.0.0
chromadb>=0.4.0
tiktoken>=0.3.0

# Modal Deployment
modal>=0.50.0